event handler done

+++ app/src/main/java/com/jacksafblaze/roomdemo/SubscriberUiState.kt
@@ -1,4 +1,4 @@
 package com.jacksafblaze.roomdemo
 
-data class UserMessage(val id: Int, val message: String)
+data class UserMessage(val id: Long, val message: String)
 data class SubscriberUiState(val messages: List<UserMessage> = emptyList())
